% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertirtemp.R
\name{celsius_a_fahrenheit}
\alias{celsius_a_fahrenheit}
\title{Funcion para convertir de Celsius a Fahtemheit}
\usage{
celsius_a_fahrenheit(celsius)
}
\arguments{
\item{celsius}{temperatura en celsius}
}
\value{
la conversion de celsius a fahrenheit
}
\description{
Funcion para convertir de Celsius a Fahtemheit
}
\examples{
temperatura_celsius <- 25
temperatura_fahrenheit <- celsius_a_fahrenheit(temperatura_celsius)
print(temperatura_fahrenheit)


}
