% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celsius_to_fahrenheit.R
\name{celsius_to_fahrenheit}
\alias{celsius_to_fahrenheit}
\title{Pasar de grados celsius a grados fahrenheit}
\usage{
celsius_to_fahrenheit(temp_centigrados)
}
\arguments{
\item{temp_centigrados}{una temperatura en grados celsius}
}
\value{
La temperatura ingresada pero en grados fahrenheit
}
\description{
La funcion \code{celsius_to_fahrenheit} toma una temperatura en celsius y la pasa a fahrenheit
}
\details{
Estos son detalles extra de la funcion
}
\examples{
celsius_to_fahrenheit(25)
celsius_to_fahrenheit(10)

}
