% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fahrenheit_a_centigrados.R
\name{fahrenheit_a_centigrados}
\alias{fahrenheit_a_centigrados}
\title{Convierte temperatura de Fahrenheit a Centigrados}
\usage{
fahrenheit_a_centigrados(temperatura_fahrenheit)
}
\arguments{
\item{temperatura_fahrenheit}{vector numérico}
}
\value{
Devuelve un vector numérico del mismo largo que \code{temperatura_fahrenheit}
}
\description{
La función acepta tanto valores numéricos como vectores numéricos.
}
\examples{
fahrenheit_a_centigrados(100)

}
