% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulgadas_a_cm.R
\name{pulgadas_a_centimetros}
\alias{pulgadas_a_centimetros}
\title{Convierte pulgadas a centímetros}
\usage{
pulgadas_a_centimetros(pulgadas)
}
\arguments{
\item{pulgadas}{Un valor o vector numérico en pulgadas}
}
\value{
Un valor o vector numérico en centímetros
}
\description{
Toma como entrada un valor o vector numérico en pulgadas y devuelve su equivalente en centímetros
}
\examples{
pulgadas_a_centimetros(10) # devuelve 25.4

}
