% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datos}
\alias{datos}
\title{Conjunto de datos de ejemplo}
\format{
los datos son generados de forma sintética.
\describe{
\item{a}{muestra de datos}
\item{b}{muestra de datos multiplicada por cinco}
}
}
\source{
generación propia con funciones de R
}
\usage{
datos
}
\description{
Los datos son utilizados como ejemplo de muestra de función de suma.
}
\keyword{datasets}
