% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejercicio1.R
\name{F_to_C}
\alias{F_to_C}
\title{Centigrados a fahrenheit}
\usage{
F_to_C(temp_fahrenheit)
}
\arguments{
\item{temp_c}{me retorna la converción de centigrados a fahrenheit}
}
\description{
Centigrados a fahrenheit
}
\examples{
(11 * 9/5) + 32

}
