% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulgadas_a_centimetros.R
\name{pulgadas_a_centimetros}
\alias{pulgadas_a_centimetros}
\title{Convierte pulgadas a centímetros}
\usage{
pulgadas_a_centimetros(medida_pulgadas)
}
\arguments{
\item{medida_pulgadas}{vector numérico}
}
\value{
Devuelve un vector numérico del mismo largo que \code{medida_pulgadas}
}
\description{
La función acepta tanto valores numéricos como vectores numéricos.
}
\examples{
pulgadas_a_centimetros(7.5)

}
