% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma.R
\name{suma}
\alias{suma}
\title{Suma dos numeros}
\usage{
suma(x, y)
}
\arguments{
\item{x}{es un vetor numerico}

\item{y}{es un vector numerico}
}
\value{
Un vector numerico con la suma de \code{x} e \code{y}
}
\description{
Esta función toma 2 numeros positivos y los suma.
1
}
\examples{
suma(2, 2)
suma(1, 10)

}
