% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma.R
\name{suma}
\alias{suma}
\title{The sum of two numbers}
\usage{
suma(x = 2, y = 2)
}
\arguments{
\item{x}{It is a positive numeric value.}

\item{y}{It is a positive numeric value.}
}
\value{
A numeric vector containing the sum of \code{x} and \code{y}.
}
\description{
This function \code{suma()} takes two numbers and returns their sum. It does not accept negative values and provides a warning if a non-numeric input is provided.
}
\details{
This are extra details of this function
}
\examples{
suma(1,1)
suma(-1,1)
suma("1",1)
suma(30,50)


}
