% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma.R
\name{suma}
\alias{suma}
\title{Suma dos numeros}
\usage{
suma(x = 0, y = 0)
}
\arguments{
\item{x}{un vector numérico positivo}

\item{y}{un vector numérico positivo}
}
\value{
Un vector numerico con la suma de \verb{x´ e }y´.
}
\description{
La función `suma()´ toma dos números positivos y los suma.
}
\details{
Estos son detalles extras de la función.
}
\examples{
suma(2, 2)
suma(1, 10)

}
