% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma.R
\name{suma}
\alias{suma}
\title{Suma dos numeros}
\usage{
suma(x = 2, y = 2)
}
\arguments{
\item{x}{un vector numérico positivo}

\item{y}{un vector numérico}
}
\value{
Un vector numérico con la suma de \code{x} e \code{y}.
}
\description{
Esta función \code{suma()} toma dos números  y los suma.
}
\details{
Estos son detalles extras de la funcion.
}
\examples{
suma(2,2)
suma(1,10)



}
