% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma.R
\name{suma}
\alias{suma}
\title{Suma dos números con validaciones básicas}
\usage{
suma(x = 3, y = 2)
}
\arguments{
\item{x}{Un número (por defecto 3).}

\item{y}{Un número (por defecto 2).}
}
\value{
\itemize{
\item Si ambos argumentos son numéricos y no negativos, devuelve la suma (\code{x + y}).
\item Si alguno es negativo, devuelve el texto \code{"No puedo sumar negativos"}.
\item Si alguno no es numérico, genera un error.
}
}
\description{
Esta función suma dos valores numéricos positivos.
Si alguno de los valores no es numérico, devuelve un error.
Si alguno de los valores es negativo, devuelve un mensaje de advertencia en lugar de realizar la suma.
}
\examples{
suma()
suma(5, 7)
suma(5, -3)
# suma("a", 4)  # Genera un error

}
