% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma.R
\name{suma}
\alias{suma}
\title{Suma dos numeros}
\usage{
suma(x = 2, y = 2)
}
\arguments{
\item{x}{un vector numrico positivo}

\item{y}{un vector numrico}
}
\value{
Un vector numerico con la suma de \code{x} e \code{y}.
}
\description{
La función \code{suma()} toma 2 números positivos y los suma.
}
\details{
Estos son detalles extras de la función.
}
\examples{
suma(2, 2)
suma(1, 10)


}
